+++ controllers/products/__init__.py
@@ -603,8 +603,7 @@ def add_product():
 
 				image.save(os.path.join('static', imagename))
 			else:
-				if permission == "true":
-					errormsg = "Please take a good photo"
+				imagename = ""
 
 			if errormsg == "":
 				product = Product(locationid, menuid, name, info, imagename, options, others, sizes, price)
@@ -656,7 +655,7 @@ def update_product():
 				oldimage = product.image
 
 				if oldimage != newimagename:
-					if oldimage != "" and os.path.exists("static/" + oldimage):
+					if oldimage != "" and oldimage != None and os.path.exists("static/" + oldimage):
 						os.remove("static/" + oldimage)
 
 					image.save(os.path.join('static', newimagename))
@@ -685,7 +684,7 @@ def remove_product(id):
 	if product != None:
 		image = product.image
 
-		if os.path.exists("static/" + image):
+		if image != "" and image != None and os.path.exists("static/" + image):
 			os.remove("static/" + image)
 
 		db.session.delete(product)
